Saving image to BMP

+++ Color.h
@@ -5,9 +5,9 @@
 
 struct Color
 {
-	uint32_t r;
-	uint32_t g;
-	uint32_t b;
+	uint8_t r;
+	uint8_t g;
+	uint8_t b;
 };
 
 #endif
